+++ src/tools/testing/sf-test-apex.ts
@@ -36,7 +36,7 @@ RunSpecifiedTests="Run the Apex tests I specify, these will be specified in the 
   classNames: z.array(z.string()).describe(
     `Apex tests classes to run.
             if Running all tests, all tests should be listed
-            if unsure, find apex classes matching the pattern *.cls, that include the @isTest decorator in the file and then join their test names together with ','
+            if unsure, find apex classes matching the pattern **/classes/*.cls, that include the @isTest decorator in the file and then join their test names together with ','
 `
   ),
   usernameOrAlias: usernameOrAliasParam,
@@ -59,8 +59,8 @@ export type ApexRunTests = z.infer<typeof runApexTestsParam>;
  */
 export const registerToolRunApexTest = (server: SfMcpServer): void => {
   server.tool(
-    'sf-run-apex-tests',
-    `Run Apex tests in an org. 
+    'sf-test-apex',
+    `Run Apex tests in an org.
 
 AGENT INSTRUCTIONS:
 If the user doesn't specify what to test, take context from the currently open file
@@ -70,10 +70,14 @@ this should be chosen when a file in the 'classes' directory is mentioned
 
 EXAMPLE USAGE:
 Run tests A, B, C.
-Run the tests, find apex classes matching the pattern *.cls, that include the @isTest decorator in the file and then join their test names together with ','
+Run the tests, find apex classes matching the pattern **/classes/*.cls, that include the @isTest decorator in the file and then join their test names together with ','
 Run all tests in the org.
 `,
     runApexTestsParam.shape,
+    {
+      title: 'Apex Tests',
+      openWorldHint: false,
+    },
     async ({ testLevel, usernameOrAlias, classNames, directory }) => {
       if (testLevel !== TestLevel.RunSpecifiedTests && classNames?.length && classNames?.length >= 1) {
         return textResponse("You can't specify which tests to run without setting testLevel='RunSpecifiedTests'", true);